+++ chains/evm/cli/utils/utils.go
@@ -1,14 +1,6 @@
 package utils
 
 import (
-	"errors"
-	gomath "math"
-	"math/big"
-	"strings"
-
-	"github.com/ChainSafe/chainbridge-core/chains/evm/calls"
-	"github.com/ChainSafe/chainbridge-core/chains/evm/calls/evmgaspricer"
-
 	"github.com/spf13/cobra"
 )
 
@@ -21,36 +13,4 @@ var UtilsCmd = &cobra.Command{
 func init() {
 	UtilsCmd.AddCommand(simulateCmd)
 	UtilsCmd.AddCommand(hashListCmd)
-}
-
-// UserAmountToWei converts decimal user friendly representation of token amount to 'Wei' representation with provided amount of decimal places
-// eg UserAmountToWei(1, 5) => 100000
-func UserAmountToWei(amount string, decimal *big.Int) (*big.Int, error) {
-	amountFloat, ok := big.NewFloat(0).SetString(amount)
-	if !ok {
-		return nil, errors.New("wrong amount format")
-	}
-	ethValueFloat := new(big.Float).Mul(amountFloat, big.NewFloat(gomath.Pow10(int(decimal.Int64()))))
-	ethValueFloatString := strings.Split(ethValueFloat.Text('f', int(decimal.Int64())), ".")
-
-	i, ok := big.NewInt(0).SetString(ethValueFloatString[0], 10)
-	if !ok {
-		return nil, errors.New(ethValueFloat.Text('f', int(decimal.Int64())))
-	}
-
-	return i, nil
-}
-
-func WeiAmountToUser(amount *big.Int, decimals *big.Int) (*big.Float, error) {
-	amountFloat, ok := big.NewFloat(0).SetString(amount.String())
-	if !ok {
-		return nil, errors.New("wrong amount format")
-	}
-	return new(big.Float).Quo(amountFloat, big.NewFloat(gomath.Pow10(int(decimals.Int64())))), nil
-}
-
-type GasPricerWithPostConfig interface {
-	calls.GasPricer
-	SetClient(client evmgaspricer.LondonGasClient)
-	SetOpts(opts *evmgaspricer.GasPricerOpts)
 }
